enums: ensure that enum values passed in match the ones passed back out
When taking enum values from Python we cast the signed values to the
enum type, which in case the enum type is unsigned leads to a defined
overflow.

When going the opposite way and creating Python enum values from
C enum values we have to do the reverse. Since casts from unsigned int
to signed int are undefined for > INT_MAX we need to do the type
conversion manually.

+++ cairo/context.c
@@ -416,8 +416,7 @@ pycairo_get_miter_limit (PycairoContext *o) {
 
 static PyObject *
 pycairo_get_operator (PycairoContext *o) {
-  /* See NOTE ['cairo_operator_t' underlying type] */
-  RETURN_INT_ENUM(Operator, (int)cairo_get_operator (o->ctx));
+  RETURN_INT_ENUM(Operator, cairo_get_operator (o->ctx));
 }
 
 static PyObject *
